Move title tag test
